+++ src/main/clojure/com/github/clojure_repl/intellij/configuration/factory/base.clj
@@ -8,7 +8,8 @@
    [com.intellij.execution.ui ConsoleView]
    [com.intellij.ide.plugins PluginManagerCore]
    [com.intellij.openapi.actionSystem AnAction]
-   [com.intellij.openapi.extensions PluginId]))
+   [com.intellij.openapi.extensions PluginId]
+   [com.intellij.openapi.project Project]))
 
 (set! *warn-on-reflection* true)
 
@@ -16,11 +17,11 @@
                               :console nil}))
 (defonce editor-view* (atom nil))
 
-(defn ^:private initial-repl-text []
-  (let [{:keys [clojure java nrepl]} (-> @db/db* :current-nrepl :versions)]
+(defn ^:private initial-repl-text [project]
+  (let [{:keys [clojure java nrepl]} (db/get-in project [:current-nrepl :versions])]
     (str (format ";; Connected to nREPL server - nrepl://%s:%s\n"
-                 (-> @db/db* :settings :nrepl-host)
-                 (-> @db/db* :settings :nrepl-port))
+                 (db/get-in project [:settings :nrepl-host])
+                 (db/get-in project [:settings :nrepl-port]))
          (format ";; Clojure REPL Intellij %s\n"
                  (.getVersion
                   (PluginManagerCore/getPlugin (PluginId/getId "com.github.clojure-repl"))))
@@ -33,6 +34,7 @@
   (reset! current-repl* {:handler nil
                          :console nil})
   (swap! current-repl* assoc :console (ui.repl/build-console
+                                       project
                                        {:on-eval (fn [code]
                                                    (nrepl/eval {:project project :code code}))}))
   (ui.repl/append-text (:console @current-repl*) loading-text)
@@ -56,16 +58,16 @@
     (createConsoleActions [_] (into-array AnAction []))
     (allowHeavyFilters [_])))
 
-(defn repl-disconnected []
+(defn repl-disconnected [^Project project]
   (ui.repl/close-console (:console @current-repl*))
   (reset! current-repl* {:handler nil
                          :console nil})
-  (swap! db/db* assoc :current-nrepl nil))
+  (db/assoc-in project [:current-nrepl] nil))
 
 (defn repl-started [project extra-initial-text]
-  (nrepl/clone-session)
+  (nrepl/clone-session project)
   (nrepl/eval {:project project :code "*ns*"})
-  (let [description (nrepl/describe)]
-    (swap! db/db* assoc-in [:current-nrepl :ops] (:ops description))
-    (swap! db/db* assoc-in [:current-nrepl :versions] (:versions description))
-    (ui.repl/set-initial-text (:console @current-repl*) (str (initial-repl-text) extra-initial-text))))
+  (let [description (nrepl/describe project)]
+    (db/assoc-in project [:current-nrepl :ops] (:ops description))
+    (db/assoc-in project [:current-nrepl :versions] (:versions description))
+    (ui.repl/set-initial-text project (:console @current-repl*) (str (initial-repl-text project) extra-initial-text))))